use interceptor strategy on user launcher tests
