+++ toolkit/authentication/src/androidTest/java/com/arcgismaps/toolkit/authentication/OAuthDefaultConfigurationTests.kt
@@ -1,7 +1,5 @@
 package com.arcgismaps.toolkit.authentication
 
-import android.content.Intent
-import android.net.Uri
 import androidx.activity.ComponentActivity
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
@@ -11,15 +9,10 @@ import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.uiautomator.UiDevice
 import com.arcgismaps.ArcGISEnvironment
 import com.arcgismaps.exceptions.OperationCancelledException
-import com.arcgismaps.httpcore.ArcGISHttpClient
-import com.arcgismaps.httpcore.Request
-import com.arcgismaps.httpcore.Response
 import com.arcgismaps.httpcore.authentication.ArcGISAuthenticationChallenge
 import com.arcgismaps.httpcore.authentication.ArcGISAuthenticationChallengeResponse
 import com.arcgismaps.httpcore.authentication.OAuthUserConfiguration
 import com.arcgismaps.httpcore.authentication.OAuthUserCredential
-import io.mockk.every
-import io.mockk.mockk
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.async
@@ -66,8 +59,7 @@ class OAuthDefaultConfigurationTests {
     @Test
     fun signInWithCredentials() = runTest {
         val response = testOAuthChallengeWithStateRestoration {
-            InstrumentationRegistry.getInstrumentation().context.startActivity(getSuccessfulRedirectIntent())
-//            composeTestRule.activity.startActivity(getRedirectIntent())
+            InstrumentationRegistry.getInstrumentation().context.startActivity(getSuccessfulRedirectIntent("kotlin-authentication-test-1://auth"))
         }.await().getOrNull()
         assert(response is ArcGISAuthenticationChallengeResponse.ContinueWithCredential)
         assert((response as ArcGISAuthenticationChallengeResponse.ContinueWithCredential).credential is OAuthUserCredential)
@@ -165,70 +157,3 @@ class OAuthDefaultConfigurationTests {
         return challengeResponse
     }
 }
-
-/**
- * A mock [ArcGISAuthenticationChallenge] for testing purposes.
- *
- * @since 200.5.0
- */
-fun makeMockArcGISAuthenticationChallenge() = mockk<ArcGISAuthenticationChallenge>().apply {
-    every { requestUrl } returns "https://arcgis.com"
-    every { cause } returns Throwable()
-}
-
-/**
- * Returns an intent that simulates a successful redirect from the OAuth server.
- *
- * @since 200.5.0
- */
-fun getSuccessfulRedirectIntent(): Intent {
-    return Intent().apply {
-        data = Uri.parse("kotlin-authentication-test-1://auth?code=12345")
-        flags = Intent.FLAG_ACTIVITY_CLEAR_TOP
-    }
-}
-
-/**
- * Intercepts requests to validate a token and returns a fake response that validates the token.
- *
- * @since 200.5.0
- */
-fun ArcGISHttpClient.Builder.interceptTokenRequests() {
-    interceptor { chain ->
-        chain.request().let { request ->
-            if (request.url.endsWith("sharing/rest/oauth2/token")) {
-                request.getFakeTokenResponse()
-            } else chain.proceed(request)
-        }
-    }
-}
-
-/**
- * Returns a fake response that validates the token.
- *
- * @since 200.5.0
- */
-fun Request.getFakeTokenResponse(): Response =
-    Response.builder().apply {
-        request(this@getFakeTokenResponse)
-        val bodyString =
-            """
-            {"access_token":"12345","expires_in":1800,"username":"username","ssl":true,"refresh_token":"67890","refresh_token_expires_in":1209599}
-            """.trimIndent()
-        body(
-            Response.Body.builder().contentType("text/plain")
-                .data(bodyString.byteInputStream(), bodyString.length.toLong()).build()
-        )
-        addHeader("cache-control", "no-cache, no-store, must-revalidate")
-        addHeader("connection", "keep-alive")
-        addHeader("content-encoding", "gzip")
-        addHeader("content-type", "text/plain;charset=utf-8")
-        addHeader("date", "Wed, 22 May 2024 15:04:50 GMT")
-        addHeader("expires", "0")
-        addHeader("pragma", "no-cache")
-        addHeader("response-status-code", "200")
-        addHeader("strict-transport-security", "max-age=31536000")
-        addHeader("transfer-encoding", "chunked")
-        addHeader("vary", "X-Esri-Authorization")
-        addHeader("x-content-type-options", "nosniff")
-    }.build()
